requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==5.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.11.0
